requirements.txt
torch
aim
Pillow
matplotlib
h5py
numpy<2
scipy
tqdm
polars
gdown
jax
equinox
optax
tyro
py-spy
